Increase call timeout in a test
There were spurious failures due gc pauses. The original timeout 6s
was insufficient as there were multiple pauses close to 5.
```
10:04:20, accumulated pauses: 1085 ms, max pause: 138 ms, pauses over 1000 ms: 0
10:04:25, accumulated pauses: 2667 ms, max pause: 2000 ms, pauses over 1000 ms: 1
10:04:30, accumulated pauses: 4911 ms, max pause: 4562 ms, pauses over 1000 ms: 1
10:04:35, accumulated pauses: 4290 ms, max pause: 2537 ms, pauses over 1000 ms: 2
10:04:40, accumulated pauses: 704 ms, max pause: 146 ms, pauses over 1000 ms: 0
10:04:45, accumulated pauses: 1287 ms, max pause: 229 ms, pauses over 1000 ms: 0
10:04:50, accumulated pauses: 505 ms, max pause: 37 ms, pauses over 1000 ms: 0
```

Increased timeout makes the test slow hence moving to SlowTest category

Fixes #7940
